refactor: 커스텀 훅을 이용하여 컴포넌트와 전역 상태 간 직접 의존 제거
